string: Fix type of position markers returned by string functions
"unsigned int" is only valid in 32-bit systems, not in 64-bit systems like
modern PCs.

The change in Controller::configureUrl() has fixed the configuration page
in the SDL port.

+++ bunjalloo/arm9/Controller.cpp
@@ -262,7 +262,7 @@ void Controller::configureUrl(const std::string & fileName)
   // search for ? - if found, send the encoding to configure
   // otherwise localFile the fileName
 
-  unsigned int position = fileName.find_first_of("?");
+  size_t position = fileName.find_first_of("?");
   if (position != string::npos)
   {
     string postedUrl = fileName.substr(position+1, fileName.length() - position - 1);
